+++ providers/ftp/src/airflow/providers/ftp/sensors/ftp.py
@@ -23,12 +23,7 @@ from collections.abc import Sequence
 from typing import TYPE_CHECKING
 
 from airflow.providers.ftp.hooks.ftp import FTPHook, FTPSHook
-from airflow.providers.ftp.version_compat import AIRFLOW_V_3_0_PLUS
-
-if AIRFLOW_V_3_0_PLUS:
-    from airflow.sdk import BaseSensorOperator
-else:
-    from airflow.sensors.base import BaseSensorOperator  # type: ignore[no-redef]
+from airflow.providers.ftp.version_compat import BaseSensorOperator
 
 if TYPE_CHECKING:
     try:
